# web/src/services/api/_internal/openapi.json
{"openapi":"3.1.0","info":{"title":"Chowbea PDF API","version":"0.1.0"},"paths":{"/pdf/compress":{"post":{"tags":["pdf"],"summary":"Queue one or more PDF files for compression","description":"Validate and store the uploads, then queue a compression job.","operationId":"compress_pdf_compress_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_compress_pdf_compress_post"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobAccepted"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/pdf/unlock":{"post":{"tags":["pdf"],"summary":"Queue a password removal job for a PDF file","description":"Validate and store the upload, then queue an unlock job.","operationId":"unlock_pdf_unlock_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_unlock_pdf_unlock_post"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobAccepted"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/pdf/lock":{"post":{"tags":["pdf"],"summary":"Queue a password protection job for a PDF file","description":"Validate and store the upload, then queue a lock job.","operationId":"lock_pdf_lock_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_lock_pdf_lock_post"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobAccepted"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/pdf/merge":{"post":{"tags":["pdf"],"summary":"Queue a merge of two or more PDF files","description":"Validate and store the uploads, then queue a merge job.","operationId":"merge_pdf_merge_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_merge_pdf_merge_post"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobAccepted"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/pdf/rotate":{"post":{"tags":["pdf"],"summary":"Queue a rotate/reorder of a PDF's pages","description":"Validate and store the upload, then queue a rotate job.","operationId":"rotate_pdf_rotate_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_rotate_pdf_rotate_post"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobAccepted"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/pdf/split":{"post":{"tags":["pdf"],"summary":"Queue a split of a PDF into page groups","description":"Validate and store the upload, then queue a split job.","operationId":"split_pdf_split_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_split_pdf_split_post"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobAccepted"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/pdf/convert":{"post":{"tags":["pdf"],"summary":"Queue a file conversion","description":"Validate the upload(s) and conversion pair, then queue a convert job.","operationId":"convert_pdf_convert_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_convert_pdf_convert_post"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobAccepted"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/jobs/{job_id}":{"get":{"tags":["jobs"],"summary":"Poll a job's status","operationId":"job_status_jobs__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobStatusResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/jobs/{job_id}/download":{"get":{"tags":["jobs"],"summary":"Download a finished job's result","operationId":"job_download_jobs__job_id__download_get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/queue":{"get":{"tags":["jobs"],"summary":"Public queue board","operationId":"queue_board_queue_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueueBoard"}}}}}}},"/health":{"get":{"tags":["meta"],"summary":"Liveness check","description":"Return a status payload used by load balancers, uptime checks, and\nbug reports (the commit identifies the running deploy).","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Health Get"}}}}}}}},"components":{"schemas":{"BoardEntry":{"properties":{"id_prefix":{"type":"string","title":"Id Prefix"},"tool":{"type":"string","title":"Tool"},"file_count":{"type":"integer","title":"File Count"},"total_bytes":{"type":"integer","title":"Total Bytes"},"created_at":{"type":"number","title":"Created At"}},"type":"object","required":["id_prefix","tool","file_count","total_bytes","created_at"],"title":"BoardEntry","description":"Anonymized public view of one job. Never includes filenames."},"Body_compress_pdf_compress_post":{"properties":{"files":{"items":{"type":"string","format":"binary"},"type":"array","title":"Files","description":"One or more PDF files to compress."},"quality":{"$ref":"#/components/schemas/CompressionQuality","description":"Compression preset; 'screen' is smallest, 'prepress' is highest quality.","default":"ebook"}},"type":"object","required":["files"],"title":"Body_compress_pdf_compress_post"},"Body_convert_pdf_convert_post":{"properties":{"files":{"items":{"type":"string","format":"binary"},"type":"array","title":"Files","description":"The file to convert (multiple images may combine into one PDF)."},"target":{"type":"string","title":"Target","description":"Target format: pdf, docx, md, html, txt, png, or jpeg."},"dpi":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Dpi","description":"Resolution for png/jpeg targets: 72, 150, or 300."}},"type":"object","required":["files","target"],"title":"Body_convert_pdf_convert_post"},"Body_lock_pdf_lock_post":{"properties":{"file":{"type":"string","format":"binary","title":"File","description":"A PDF to password-protect."},"password":{"type":"string","title":"Password","description":"The password that will be required to open the PDF."},"allow_printing":{"type":"boolean","title":"Allow Printing","description":"Permit printing the locked PDF.","default":true},"allow_copying":{"type":"boolean","title":"Allow Copying","description":"Permit copying/extracting text from the PDF.","default":false},"allow_editing":{"type":"boolean","title":"Allow Editing","description":"Permit editing and annotating the PDF.","default":false},"encryption":{"$ref":"#/components/schemas/EncryptionLevel","description":"Encryption strength; 'aes-256' is strongest.","default":"aes-256"}},"type":"object","required":["file","password"],"title":"Body_lock_pdf_lock_post"},"Body_merge_pdf_merge_post":{"properties":{"files":{"items":{"type":"string","format":"binary"},"type":"array","title":"Files","description":"Two or more PDF files, in merge order."}},"type":"object","required":["files"],"title":"Body_merge_pdf_merge_post"},"Body_rotate_pdf_rotate_post":{"properties":{"file":{"type":"string","format":"binary","title":"File","description":"The PDF whose pages to rotate/reorder."},"pages":{"type":"string","title":"Pages","description":"JSON list of {\"index\", \"rotation\"}; array order is the new page order."}},"type":"object","required":["file","pages"],"title":"Body_rotate_pdf_rotate_post"},"Body_split_pdf_split_post":{"properties":{"file":{"type":"string","format":"binary","title":"File","description":"The PDF to split."},"parts":{"type":"string","title":"Parts","description":"JSON list of {\"pages\": [int, ...]}; each entry becomes one output PDF."}},"type":"object","required":["file","parts"],"title":"Body_split_pdf_split_post"},"Body_unlock_pdf_unlock_post":{"properties":{"file":{"type":"string","format":"binary","title":"File","description":"A password-protected PDF to unlock."},"password":{"type":"string","title":"Password","description":"The password that opens the PDF."}},"type":"object","required":["file","password"],"title":"Body_unlock_pdf_unlock_post"},"CompressionQuality":{"type":"string","enum":["screen","ebook","printer","prepress"],"title":"CompressionQuality","description":"Quality presets that map directly to Ghostscript's -dPDFSETTINGS values.\n\nLower quality means smaller files: `screen` is the most aggressive,\n`prepress` preserves the most detail."},"EncryptionLevel":{"type":"string","enum":["aes-128","aes-256"],"title":"EncryptionLevel","description":"Encryption strength presets exposed by the API."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"JobAccepted":{"properties":{"job_id":{"type":"string","title":"Job Id"},"position":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Position"},"queue_size":{"type":"integer","title":"Queue Size"}},"type":"object","required":["job_id","position","queue_size"],"title":"JobAccepted"},"JobStatus":{"type":"string","enum":["queued","processing","done","failed"],"title":"JobStatus"},"JobStatusResponse":{"properties":{"id":{"type":"string","title":"Id"},"tool":{"type":"string","title":"Tool"},"status":{"$ref":"#/components/schemas/JobStatus"},"position":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Position"},"queue_size":{"type":"integer","title":"Queue Size"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"},"file_count":{"type":"integer","title":"File Count"},"total_bytes":{"type":"integer","title":"Total Bytes"},"created_at":{"type":"number","title":"Created At"}},"type":"object","required":["id","tool","status","position","queue_size","error","file_count","total_bytes","created_at"],"title":"JobStatusResponse"},"QueueBoard":{"properties":{"concurrency":{"type":"integer","title":"Concurrency"},"jobs_completed":{"type":"integer","title":"Jobs Completed"},"processing":{"items":{"$ref":"#/components/schemas/BoardEntry"},"type":"array","title":"Processing"},"waiting":{"items":{"$ref":"#/components/schemas/BoardEntry"},"type":"array","title":"Waiting"}},"type":"object","required":["concurrency","jobs_completed","processing","waiting"],"title":"QueueBoard"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}